requirements.txt
elasticsearch=>5.4.0
kafka-python=>1.3.3
nltk=>3.2.4
